ansible: remove deprecated config and deployment
Scale 21x will mark the second year that this code was no longer
utilized. It seems like the appropriate time to remove this deprecated
part of the scale-network repo.

A big thank you goes out to Kyle for building out this out prior to
Scale 16x. Its served us well.

Onward to the nix iteration!